ml_project/requirements.txt
numpy==1.21.5,
click==8.0.4,
scikit-learn==1.0.2,
dataclasses==0.8,
pyyaml==6.0,
marshmallow-dataclass==8.5.8,
pandas==1.3.5,
catboost==0.26.1,
pytest==7.1.2